+++ seq2seq.py
@@ -4,106 +4,12 @@
 import tensorflow as tf
 import tensorlayer as tl
 import numpy as np
-from tensorlayer.models import Model
-from tensorlayer.layers import Dense, Dropout, Input
-from tensorlayer.layers.core import Layer
-from cross_entropy_seq_loss import *
+from loss import cross_entropy_seq
 from tqdm import tqdm
 from sklearn.utils import shuffle
 from data.twitter import data
+from model_seq2seq import Seq2seq
 
-
-class Seq2seq(Model):
-    def __init__(
-            self,
-            batch_size,
-            cell_dec,
-            cell_enc,
-            embedding_layer=None,
-            is_train=True,
-            name="seq2seq"
-    ):
-        super(Seq2seq, self).__init__(name=name)
-        self.embedding_layer = embedding_layer
-        self.vocabulary_size = embedding_layer.vocabulary_size
-        self.embedding_size = embedding_layer.embedding_size
-        self.encoding_layer = tl.layers.RNN(cell=cell_enc, in_channels=self.embedding_size, return_state=True)
-        self.decoding_layer = tl.layers.RNN(cell=cell_dec, in_channels=self.embedding_size)
-        self.reshape_layer = tl.layers.Reshape([-1, cell_dec.units])
-        self.dense_layer = tl.layers.Dense(n_units=self.vocabulary_size, act=tf.nn.softmax, in_channels=cell_dec.units)
-        self.reshape_layer_after = tl.layers.Reshape([batch_size, -1, self.vocabulary_size])
-        self.reshape_layer_individual_sequence = tl.layers.Reshape([-1, 1, self.vocabulary_size])
-        
-
-    def inference(self, encoding, seq_length, start_token):
-
-        # after embedding the encoding sequence, start the encoding_RNN, then transfer the state to decoing_RNN
-        after_embedding_encoding = self.embedding_layer(encoding)
-
-        enc_rnn_ouput, state = self.encoding_layer(after_embedding_encoding, return_state=True)
-
-        
-        # for the start_token, first create a batch of it, get[Batchsize, 1]. 
-        # then embbeding, get[Batchsize, 1, embeddingsize]
-        # then RNN, get[Batchsize, 1, RNN_units]
-        # then reshape, get[Batchsize*1, RNN_units]
-        # then dense, get[Batchsize*1, vocabulary_size]
-        # then reshape, get[Batchsize, 1, vocabulary_size]
-        # finally, get Argmax of the last dimension, get next_sequence[Batchsize, 1]
-        # this next_sequence will repeat above procedure for the sequence_length time
-
-
-        batch_size = len(encoding)
-        decoding = [[start_token] for i in range(batch_size)]
-        decoding = np.array(decoding)
-        
-        after_embedding_decoding = self.embedding_layer(decoding)
-
-        feed_output, state = self.decoding_layer(after_embedding_decoding, initial_state=state, return_state=True)
-        feed_output = self.reshape_layer(feed_output)
-        feed_output = self.dense_layer(feed_output)
-        feed_output = self.reshape_layer_individual_sequence(feed_output)
-        feed_output = tf.argmax(feed_output, -1)
-        
-        final_output = feed_output
-        
-        for i in range(seq_length-1):
-            feed_output = self.embedding_layer(feed_output)
-            feed_output, state = self.decoding_layer(feed_output, state, return_state=True)
-            feed_output = self.reshape_layer(feed_output)
-            feed_output = self.dense_layer(feed_output)
-            feed_output = self.reshape_layer_individual_sequence(feed_output)
-            feed_output = tf.argmax(feed_output, -1)
-            final_output = tf.concat([final_output,feed_output], 1)
-
-
-        return final_output, state
-
-
-    def forward(self, inputs, seq_length=8, start_token=None, return_state=False):
-
-
-        if (self.is_train):
-            encoding = inputs[0]
-            after_embedding_encoding = self.embedding_layer(encoding)
-            decoding = inputs[1]
-            after_embedding_decoding = self.embedding_layer(decoding)
-            enc_rnn_output, state = self.encoding_layer(after_embedding_encoding, return_state=True)
-            
-            dec_rnn_output, state = self.decoding_layer(after_embedding_decoding, initial_state=state, return_state=True)
-            dec_output = self.reshape_layer(dec_rnn_output)
-            denser_output = self.dense_layer(dec_output)
-
-            ## reshape into [Batch_size, seq_step, vocabulary_size]
-            output = self.reshape_layer_after(denser_output)
-        else:
-            encoding = inputs
-            output, state = self.inference(encoding, seq_length, start_token)
-
-        if (return_state):
-            return output, state
-        else:
-            return output
 
 def initial_setup(data_corpus):
     metadata, idx_q, idx_a = data.load_data(PATH='data/{}/'.format(data_corpus))
@@ -149,21 +55,10 @@ if __name__ == "__main__":
 
     src_vocab_size = tgt_vocab_size = src_vocab_size + 2
 
-    """ A data for Seq2Seq should look like this:
-    input_seqs : ['how', 'are', 'you', '<PAD_ID'>]
-    decode_seqs : ['<START_ID>', 'I', 'am', 'fine', '<PAD_ID'>]
-    target_seqs : ['I', 'am', 'fine', '<END_ID>', '<PAD_ID'>]
-    target_mask : [1, 1, 1, 1, 0]
-    """
-
     num_epochs = 5
     vocabulary_size = src_vocab_size
     batch_size = 32
 
-    # seq_step_input = 3  
-    # encoding = np.random.randint(low=0,high=vocabulary_size,size=(batch_size,seq_step_input))
-    # decoding = np.random.randint(low=0,high=vocabulary_size,size=(batch_size,seq_step_output))
-    # target = np.random.randint(low=0,high=10,size=(batch_size,seq_step_output))
 
     def inference(seed):
         model_.eval()
